+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.5 - 2022-04-01
 
 - Support passing headers to request for loading font files and images via URL adresses